Exercises 4.15. - 4.23.

+++ part_4/controllers/blogs.js
@@ -1,35 +1,49 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const { userExtractor } = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
- 
+  const blogs = await Blog
+    .find({})
+    .populate('user', { name: 1, username: 1 })
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response, next) => {
+blogsRouter.post('/', userExtractor, async (request, response, next) => {
   const body = request.body
-
+  const user = request.user
+  
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes || 0,
+    user: user._id
   })
 
   const result = await blog.save()
+  user.blogs = user.blogs.concat(result._id)
+  await user.save()
+
   response.status(201).json(result)
 })
 
-blogsRouter.delete('/:id', async(request, response, next) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', userExtractor, async(request, response, next) => {
+  const blogToDelete = await Blog.findById(request.params.id)
+ 
+  const user = request.user
+ 
+  if(blogToDelete.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'Unauthorized to delete blog' })
+  }
 })
 
 blogsRouter.put('/:id', async(request, response, next) => {
   const body = request.body
 
-  console.log(body)
   const blog = {
     likes: body.likes,
   }
